Cast batch_offset to long (not int) for future-proofing
``batch_offset`` in ``generate_bulk_inserts`` ranges over
``[0, len(batch_infos))`` and indexes into a ``case_when_chain``
per-batch expression list.  At current ``_auto_chunk_size`` values
(~365 batches at the 20M-rows-per-chunk target) the int32 cast is
safe, but:

- ``_batch_id`` is long everywhere else in the engine (single-batch
  / fused / oracle / CDC schema) -- the bulk path was the lone
  outlier.
- A future bump of ``_AUTO_CHUNK_TARGET_ROWS`` or an explicit large
  ``chunk_size`` could silently overflow int32 with no assertion or
  invariant comment flagging the ceiling.

Switch to ``.cast("long")`` and add a comment linking to the
chunk-size invariant.  No runtime cost; closes the future-proofing
gap surfaced in the last review.

+++ dbldatagen/core/engine/cdc/bulk.py
@@ -76,7 +76,15 @@ def generate_bulk_inserts(
     df, raw_id = create_range_df(spark, total_inserts)
 
     local_id = (raw_id % F.lit(inserts_per_batch)).cast("long")
-    batch_offset = F.floor(raw_id.cast("double") / F.lit(inserts_per_batch)).cast("int")
+    # ``batch_offset`` is in ``[0, len(batch_infos))`` and indexes into
+    # the ``case_when_chain`` per-batch expression list.  Use long (not
+    # int) to stay consistent with ``_batch_id`` long everywhere else
+    # in the engine -- chunk sizes bounded by ``_auto_chunk_size``
+    # (~365 at 20M rows/chunk) don't hit int32 today, but a future
+    # ``_AUTO_CHUNK_TARGET_ROWS`` bump or an explicit large
+    # ``chunk_size`` could silently overflow.  Long closes the window
+    # without cost.
+    batch_offset = F.floor(raw_id.cast("double") / F.lit(inserts_per_batch)).cast("long")
 
     pk_cols = get_pk_columns(table_spec)
